Some changes are made
